use double value everywhere to compute chominance adjustement. Clamp adjusted value with noise between 0.0 and 1.0. CCBUGS: 148540
svn path=/trunk/extragear/graphics/digikam/; revision=1106666

+++ libs/dimg/filters/fx/filmgrainfilter.h
@@ -94,7 +94,7 @@ private:
     int    randomize  (int value, bool sixteenbit, int range);
 
     // Chrominance methods.
-    double randomizeChroma(double value, bool sixteenbit, int range);
+    double randomizeChroma(double value, bool sixteenbit, double range);
 
 private:
